packaging/requirements-pack.txt
pyinstaller==6.17.0
numpy
pandas
matplotlib
scipy
openpyxl
requests
openai
pywebview
